requirements.txt
Django
gunicorn
psycopg2-binary
dj-database-url
python-decouple
whitenoise
